src/frontend/pyproject.toml
[tool.poetry]
name = "streaming-explorer-frontend"
version = "0.1.0"
description = "A comprehensive exploration of shows available on popular streaming platforms."
authors = ["Omar <omar120cuet@gmail.com>"]
license = "MIT"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
streamlit = "1.29.0"
plotly = "5.18.0"
scikit-learn = "1.3.2"
python-dotenv = "1.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"